back/requirements.txt
fastapi
uvicorn
playwright
spacy
pydantic
apify-client
python-dotenv